providers: rename _helpers module to _logs
