+++ charmcraft/providers/__init__.py
@@ -18,6 +18,6 @@
 
 from ._buildd import CharmcraftBuilddBaseConfiguration  # noqa: F401
 from ._get_provider import get_provider  # noqa: F401
-from ._helpers import capture_logs_from_instance  # noqa: F401
+from ._logs import capture_logs_from_instance  # noqa: F401
 from ._lxd import LXDProvider  # noqa: F401
 from ._provider import Provider  # noqa: F401
